Conflicts:
	src/java/MainTests/registerTest.java
	src/java/SqlConnection/DBConfig.properties
	src/java/SqlConnection/SqlConnection.java

+++ src/java/Dao/UserDao.java
@@ -22,12 +22,12 @@ import java.util.ArrayList;
 public class UserDao implements UserDAOInterface {
 
     private SqlConnection sql = null;
-    
-    public UserDao(SqlConnection sql){
+
+    public UserDao(SqlConnection sql) {
         this.sql = sql;
     }
-    
-    public UserDao(){
+
+    public UserDao() {
         this.sql = new SqlConnection();
     }
 
@@ -64,7 +64,6 @@ public class UserDao implements UserDAOInterface {
             //System.out.println(email);
             if (CheckUserExistsByEmail(email)) {
                 User u = getUserByEmail(email);
-                System.out.println(u.getEmail());
                 return BCrypt.checkpw(hash, u.getUserHash());
             } else {
                 System.out.println("didn't exist");
@@ -146,14 +145,14 @@ public class UserDao implements UserDAOInterface {
     @Override
     public boolean CheckUserExistsByEmail(String email) {
         try {
-            sql.setPs(sql.getConn().prepareStatement("select * from users where email=?"));
+            sql.setPs(sql.getConn().prepareStatement("select * from users WHERE email=?"));
             sql.getPs().setString(1, email);
 
             ResultSet rst;
             rst = sql.getPs().executeQuery();
 
             if (rst.next()) {
-                System.out.println("user exists");
+                System.out.println("user exi");
                 return true;
             }
             System.out.println("user doesn't exist yet");
@@ -174,19 +173,49 @@ public class UserDao implements UserDAOInterface {
     public boolean updateUser(User user
     ) {
         try {
-            sql.setPs(sql.getConn().prepareStatement("UPDATE users SET user_fullname=?,question=?,answer_hash=?, has_disabled_badge=? WHERE email = ?"));
+            if (CheckUserExistsByEmail(user.getEmail())) {
+                sql.setPs(sql.getConn().prepareStatement("UPDATE users SET user_fullname=?, user_type=?,question=?,answer_hash=?, has_disabled_badge=? WHERE email = ?"));
 
-            sql.getPs().setString(1, user.getUserFullname());
-            sql.getPs().setString(2, user.getQuestion());
-            sql.getPs().setString(3, user.getAnswer_hash());
-            sql.getPs().setBoolean(4, user.getHasDisabledBadge());
-            sql.getPs().setString(5, user.getEmail());
+                sql.getPs().setString(1, user.getUserFullname());
+                
+                sql.getPs().setString(2, user.getUserType());
+                
+                sql.getPs().setString(3, user.getQuestion());
+                sql.getPs().setString(4, user.getAnswer_hash());
+                sql.getPs().setBoolean(5, user.getHasDisabledBadge());
+                
+                sql.getPs().setString(6, user.getEmail());
 
-            sql.getPs().executeUpdate();
-
-            return true;
+                sql.getPs().executeUpdate();
+                return true;
+            } else {
+                return false;
+            }
 
         } catch (SQLException se) {
+            System.out.println("SQL Exception occurred: " + se.getMessage());
+            se.printStackTrace();
+            return false;
+        } catch (Exception e) {
+            System.out.println("Exception occurred: " + e.getMessage());
+            e.printStackTrace();
+            return false;
+        }
+    }
+    
+    @Override
+    public boolean deleteUser(int uid) {
+        try {
+                sql.setPs(sql.getConn().prepareStatement("DELETE from users WHERE user_id = ?"));
+
+                sql.getPs().setInt(1, uid);
+
+
+                sql.getPs().executeUpdate();
+                return true;
+            }
+
+         catch (SQLException se) {
             System.out.println("SQL Exception occurred: " + se.getMessage());
             se.printStackTrace();
             return false;
@@ -206,12 +235,11 @@ public class UserDao implements UserDAOInterface {
             // Execute the query
             rst = sql.getPs().executeQuery();
             User user = null;
-            while (rst.next()) {
+            if (rst.next()) {
                 user = new User(rst.getInt("user_id"), rst.getString("user_fullname"), rst.getString("email"),
                         rst.getString("hash"), rst.getString("user_type"), rst.getString("question"),
                         rst.getString("answer_hash"), rst.getBoolean("has_disabled_badge"));
             }
-            System.out.println(user.toString());
 
             return user;
         } catch (SQLException se) {
@@ -301,7 +329,7 @@ public class UserDao implements UserDAOInterface {
                 sql.getPs().setString(1, HashedSaltedPw[0]);
                 sql.getPs().setInt(2, u.getUserNo());
                 sql.getPs().executeUpdate();
-                
+
                 return true;
 
             } else {
@@ -322,20 +350,23 @@ public class UserDao implements UserDAOInterface {
     @Override
     public boolean AdminDeletesUser(User user) {
         try {
-            sql.setPs(sql.getConn().prepareStatement("DELETE FROM salt WHERE user_id = ?"));
+            if (!CheckUserExistsByEmail(user.getEmail())) {
+                sql.setPs(sql.getConn().prepareStatement("DELETE FROM salt WHERE user_id = ?"));
 
-            sql.getPs().setInt(1, user.getUserNo());
+                sql.getPs().setInt(1, user.getUserNo());
 
-            sql.getPs().executeUpdate();            
-            
+                sql.getPs().executeUpdate();
 
-            sql.setPs(sql.getConn().prepareStatement("DELETE FROM users WHERE user_id = ?"));
+                sql.setPs(sql.getConn().prepareStatement("DELETE FROM users WHERE user_id = ?"));
 
-            sql.getPs().setInt(1, user.getUserNo());
+                sql.getPs().setInt(1, user.getUserNo());
 
-            sql.getPs().executeUpdate();
+                sql.getPs().executeUpdate();
 
-            return true;
+                return true;
+            } else {
+                return false;
+            }
 
         } catch (SQLException se) {
             System.out.println("SQL Exception occurred: " + se.getMessage());
@@ -347,10 +378,11 @@ public class UserDao implements UserDAOInterface {
             return false;
         }
     }
+
     @Override
     public boolean AdminUpdatesUserTypes(User user) {
         try {
-
+            if (CheckUserExistsByEmail(user.getEmail())) {
                 sql.setPs(sql.getConn().prepareStatement("UPDATE users SET user_type = ? WHERE email = ?"));
 
                 sql.getPs().setString(1, user.getUserType());
@@ -358,7 +390,10 @@ public class UserDao implements UserDAOInterface {
 
                 sql.getPs().executeUpdate();
 
-            return true;
+                return true;
+            } else {
+                return false;
+            }
 
         } catch (SQLException se) {
             System.out.println("SQL Exception occurred: " + se.getMessage());
@@ -369,5 +404,5 @@ public class UserDao implements UserDAOInterface {
             e.printStackTrace();
             return false;
         }
-    }    
+    }
 }
